Move `logging` to from `pipeline` to `utils/logging.py`
